feat: 小数类型由double改为decimal，确保计算精确。扩充nativelib

+++ Lang/RtokenKit.cs
@@ -38,7 +38,7 @@ namespace RML.Lang {
             }
 
             if (StrKit.IsNumberStr(str) == 1) {
-                return new Rtoken(Rtype.Float, Convert.ToDouble(str));
+                return new Rtoken(Rtype.Float, Convert.ToDecimal(str));
             }
 
             if (str.StartsWith('"')) {
